Update README.md
Adding a new template for Project Posts. Project Posts will be for ongoing web3 projects to post about their impact and to replay basic understanding and overview to the community at large. They are intended to be a place for learning rather than marketing, and to create a thread of project documentation in web3.